fetch/metadata/generated/header-refresh.https.optional.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/header-refresh.optional.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for HTTP "Refresh" header</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, test) {
    const win = window.open();
    test.add_cleanup(() => win.close());

    win.location = `/common/refresh.py?location=${encodeURIComponent(url)}`

    return new Promise((resolve) => {
      addEventListener('message', (event) => {
          if (event.source === win) {
            resolve();
          }
        });
    });
  }

  const responseParams = {
    mime: 'text/html',
    body: `<script>opener.postMessage(0, '*')</${''}script>`
  };

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsOrigin'], responseParams), t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-origin']);
        });
  }, 'sec-fetch-frame-ancestors - Same origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsCrossSite'], responseParams), t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['cross-site']);
        });
  }, 'sec-fetch-frame-ancestors - Cross-site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(
        makeRequestURL(key, ['httpsSameSite'], responseParams), t
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-site']);
        });
  }, 'sec-fetch-frame-ancestors - Same site');
  </script>
  </body>
</html>
